fix(comment): 관리자 라우터 위로 올림

See merge request sw_track/class_03/web_project_2/team17/bob-hub-server!23

+++ src/services/comment-service.js
@@ -1,4 +1,5 @@
 const { commentModel } = require("../db/models");
+const buildRes = require("../util/build-response");
 
 class CommentService {
   constructor(commentModel) {
@@ -7,27 +8,43 @@ class CommentService {
 
   async create(commentDTO) {
     const result = await this.commentModel.create(commentDTO);
-    return result;
+    return buildRes("c", result);
   }
 
-  async getAll() {
+  async getAllByAdmin() {
     const comments = await this.commentModel.getAll();
+
     return comments;
   }
 
   async getByShopId(shopId) {
-    const comments = await this.commentModel.getById(shopId);
+    const comments = await this.commentModel.getByShopId(shopId);
     return comments;
   }
 
-  async update(newCommentDTO, commentId) {
-    const comment = await this.commentModel.update(newCommentDTO, { commentId });
+  async updateByAuth(newCommentDTO, commentId) {
+    const { userId } = newCommentDTO;
 
-    return comment;
+    const currentComment = await this.commentModel.getByCommentId(commentId);
+    const isByAuth = currentComment.userId === userId;
+    if (!isByAuth) throw new Error("권한이 없습니다.");
+
+    const result = await this.commentModel.update(newCommentDTO, { commentId });
+    return buildRes("u", result);
   }
 
-  async deleteById(commentId) {
-    await commentModel.deleteById(commentId);
+  async deleteByAuth(userId, commentId) {
+    const currentComment = await this.commentModel.getByShopId(commentId);
+    const isByAuth = currentComment.userId === userId;
+    if (!isByAuth) throw new Error("권한이 없습니다.");
+
+    const result = await commentModel.deleteById(commentId);
+    return buildRes("d", result);
+  }
+
+  async deleteByAdmin(commentId) {
+    const result = await commentModel.deleteById(commentId);
+    return buildRes("d", result);
   }
 }
